use optLong to reading "maxWaitTime" attribute from json

+++ src/org/hyperic/hq/events/server/session/Escalation.java
@@ -110,7 +110,7 @@ public class Escalation extends PersistedObject
         }
         setName(json.getString("name"));
         setAllowPause(json.optBoolean("allowPause"));
-        setMaxWaitTime(json.optInt("maxWaitTime"));
+        setMaxWaitTime(json.optLong("maxWaitTime"));
         setNotifyAll(json.optBoolean("notifyAll"));
         setActions(
             getEscalationActions(json.getJSONArray("actions")));
